Update sip2-ruby to 0.1.2, specify ruby version and add dev env
Ruby versions after 3.3 have an issue in nixpkgs where default gems are
reported as ignored due to missing native extensions. We patch bundler
and use the patched bundler in our wanted ruby version, and override
bundlerApp to use this version ...

To ease development like this we add a devShell to the flake and include
the patched ruby and also bundix for regenerating gemset.nix.