Move each package file to a subdir
Use the nix convention of putting each package in its own subdir with a
default.nix file.

Also replace the meta mkCallPackage function and just use callPackage
directly. This is (just slightly) more verbose but makes it clear what
is going on and should be more flexible if some package requires special
treatment.