Parameterize posixTools and javaRuntime for handle-server
The old closure pulled in a full JDK with gui support, and also the full
coreutils for a simple "ls". We now accept a java runtime as
javaRuntime, and the desired posix tools as posixTools. The former
defaults to jdk_headless, while the latter defaults to coreutils on
macOS and busybox on Linux.